# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SatL Engine API
  description: |-
    The Docker Engine REST API v1.43 as `satld` implements it -- the only external surface SatL has (invariant #8). Every intentional divergence from Docker carries a numbered entry in `docs/api-compat.md`, cited inline throughout this document.

    The daemon listens on the unix socket `/var/run/satl.sock`. There is no TCP listener for this API: the cluster's own protocol (`satl.internal.v1`) is a separate, node-to-node-only surface on ports 2377 and 2378.

    ## Version prefixes

    Every operation is reachable both on its bare path (`/info`) and under a Docker version prefix (`/v1.43/info`). Prefixes are accepted from 1.24 up to 1.43; a prefix outside that range is a `400` in Docker's error shape, not a 404. A leading segment that merely starts with `v` but is not `v<major>.<minor>` -- `/v1/info` -- is not version negotiation at all and routes (and 404s) as an ordinary path, where Docker would parse `1` as a version.

    ## Responses this document cannot attach to an operation

    - **Unknown path**: `404` with the body `{"message": "page not found"}`. Every Docker endpoint SatL does not implement answers this shape -- attach, commit, export, rename, restart, pause, update, top, stats, changes, archive, build, push, and the rest of the `/images/{name}/*` family (api-compat #22, #144, #152).
...
    - **Out-of-range version prefix**: `400` with Docker's error body, minted before routing.

...

...

    ## Error shape
...
servers:
...
paths:
...
  /images/create:
    post:
      tags:
      - Image
      summary: '`POST /images/create?fromImage=&tag=&platform=`.'
      description: '**Newline-delimited JSON, not a JSON array.** The response is `200` immediately, then one Docker `JSONMessage` per line (each terminated by `\r\n`) as the pull proceeds -- including the failure, which is reported as an `error` line on an already-successful response, exactly like Docker. The schema below describes one such line. `fromSrc` (import) is a 501, and `X-Registry-Auth` is accepted in base64url *or* standard base64 (api-compat #16).'
      operationId: ImageCreate
      parameters:
      - name: fromImage
        in: query
        description: Image reference to pull.
        required: false
        schema:
          type: string
      - name: tag
        in: query
        description: Tag or digest, when `fromImage` carries none.
        required: false
        schema:
          type: string
      - name: platform
        in: query
        description: '`os/arch[/variant]` (api-compat #9).'
        required: false
        schema:
          type: string
      - name: fromSrc
        in: query
        description: 'Rejected with 501: importing from a source is not supported (api-compat #16).'
        required: false
        schema:
          type: string
      - name: X-Registry-Auth
        in: header
        description: 'base64url or standard-base64 `AuthConfig` document (api-compat #16).'
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: One JSON progress object per line, not a JSON array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonMessage'
        '400':
          description: Unparsable reference, platform or `X-Registry-Auth`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: No such image in the registry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: '`?fromSrc=` was set, or the daemon has no image store wired.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /images/json:
    get:
      tags:
      - Image
      summary: '`GET /images/json`.'
      description: 'Images this node holds. `all`, `filters`, `digests` and `shared-size` are not read; each row carries an extra `Platform`, `SharedSize` is always 0, `Labels` always null and `ParentId` always empty (api-compat #15). Image content is node-local (api-compat #130).'
      operationId: ImageList
      responses:
        '200':
          description: One row per image.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageSummaryResponse'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /images/{name}:
    delete:
      tags:
      - Image
      summary: '`DELETE /images/{name}?force=&noprune=`.'
      description: 'Removes one reference, and the content it was the last reference to. The response is Docker''s bare array, which leaves nowhere for what the layer sweep deferred, so that count travels in the SatL-only `X-Satl-Deferred-Layers` response header (api-compat 156). Served through the `/images/{name}/*` tail wildcard: a tail whose last segment is `tag`, `json`, `push`, `history`, `get` or `load` is read as a verb, so a repository whose last component is literally one of those words has to be spelled with its tag.'
      operationId: ImageDelete
      parameters:
      - name: name
        in: path
        description: Image reference; may carry slashes.
        required: true
        schema:
          type: string
      - name: force
        in: query
        description: 'Override a refusal that says `must force`: a reference held only by terminal tasks, or an image reachable from several references. It never overrides a live claim (api-compat 161). Docker `BoolValue` semantics.'
        required: false
        schema:
          type: string
      - name: noprune
        in: query
        description: Skip the layer and content sweep, and with it the two agreeing passes it takes. Docker's meaning -- keep untagged *parent* images -- has no analogue here, because SatL images have no parents (api-compat 155, 157). Docker `BoolValue` semantics.
        required: false
        schema:
          type: string
      responses:
        '200':
          description: What was untagged and what was deleted.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageDeleteResponseItem'
        '404':
          description: No such image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: 'Something still claims the image. Two arms: `(cannot be forced)` when a non-terminal task or a service that still wants tasks names it -- `force` does not override that -- and `(must force)` when only terminal tasks do, or when several references resolve to the same image (api-compat 159, 161).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /images/{name}/json:
    get:
      tags:
      - Image
      summary: '`GET /images/{name}/json` — Docker''s `ImageInspect`.'
      description: 'Served through the `/images/{name}/*` tail wildcard, because an image name may carry slashes. The rest of that family -- `history`, `get`, `load` and `push` -- stays 404 `page not found` (api-compat #22, #152).'
      operationId: ImageInspect
      parameters:
      - name: name
        in: path
        description: Image reference; may carry slashes.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The image document.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageInspectResponse'
        '404':
          description: No such image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /images/{name}/tag:
    post:
      tags:
      - Image
      summary: '`POST /images/{name}/tag?repo=&tag=` — Docker''s exact route.'
      description: 'Adds one more reference to the same store entry: no blob is copied, both names keep resolving, and tagging an image with the name it already has is a no-op success (api-compat #22). Served through the `/images/{name}/*` tail wildcard, because an image name may carry slashes.'
      operationId: ImageTag
      parameters:
      - name: name
        in: path
        description: Source image reference; may carry slashes (`ghcr.io/team/app:v1`).
        required: true
        schema:
          type: string
      - name: repo
        in: query
        description: Target repository.
        required: false
        schema:
          type: string
      - name: tag
        in: query
        description: Target tag; `latest` when omitted.
        required: false
        schema:
          type: string
      responses:
        '201':
          description: Tagged.
        '400':
          description: The target reference is unparsable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: No such source image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
components:
  schemas:
...
    ErrorBody:
      type: object
      description: 'Docker error envelope: every non-2xx response body is `{"message": "..."}`.'
      required:
      - message
      properties:
        message:
          type: string
          description: Human-readable error message.
...
    ImageDeleteResponseItem:
      type: object
      description: |-
        One entry of `ImagesDeleted`. Exactly one field is ever set, which is
        Docker's shape.
      properties:
        Deleted:
          type:
          - string
          - 'null'
          description: Content that was deleted.
        Untagged:
          type:
          - string
          - 'null'
          description: A reference that stopped pointing at an image.
    ImageGraphDriver:
      type: object
      description: The `GraphDriver` block of an [`ImageInspectResponse`].
      required:
      - Name
      properties:
        Data:
          type:
          - object
          - 'null'
          description: Always `null`.
          additionalProperties:
            type: string
          propertyNames:
            type: string
        Name:
          type: string
          description: 'Always `zfs` (invariant #5).'
...
    ImageInspectResponse:
      type: object
      description: |-
        `GET /images/{name}/json` — Docker's `ImageInspect`.

        Aggregated by image ID, where [`ImageSummaryResponse`] is one row per
        stored reference (api-compat 160). Several fields are structurally empty
        rather than absent, because Docker clients read them positionally:
        `Parent`, `Comment`, `Author` and `DockerVersion` have no source in an OCI
        pull (api-compat #41), and `Config.Labels` is `null` because the image
        config parser does not read a `Labels` block.
      required:
      - Id
      - RepoTags
      - RepoDigests
      - Parent
      - Comment
      - Created
      - Author
      - Config
      - Architecture
      - Os
      - Size
      - VirtualSize
      - GraphDriver
      - RootFS
      properties:
        Architecture:
          type: string
          description: Architecture (`amd64`).
        Author:
          type: string
          description: Always empty.
        Comment:
          type: string
          description: Always empty.
        Config:
          $ref: '#/components/schemas/ImageInspectConfig'
          description: The runnable configuration.
        Created:
          type: string
          description: |-
            Creation time, RFC 3339 with nanoseconds; the zero time when the image
            config carried none.
        GraphDriver:
          $ref: '#/components/schemas/ImageGraphDriver'
          description: Storage driver block; always the ZFS driver with no data.
        Id:
          type: string
          description: Image ID (`sha256:…`) — SatL's is the manifest digest.
        Os:
          type: string
          description: Operating system (`freebsd`).
        Parent:
          type: string
          description: 'Parent image ID; always empty (api-compat #41).'
        Platform:
          type:
          - string
          - 'null'
          description: Image platform, `os/arch` — **SatL extension**, as on the summary.
        RepoDigests:
          type: array
          items:
            type: string
          description: Every `repository@digest` resolving to this image.
        RepoTags:
          type: array
          items:
            type: string
          description: Every `repository:tag` resolving to this image.
        RootFS:
          $ref: '#/components/schemas/ImageRootFs'
          description: |-
            The rootfs layer list. Docker spells this `RootFS`, which `PascalCase`
            would render `RootFs`, so it is renamed explicitly — the same class of
            irregular spelling as `ImageID` and `OSType` elsewhere in this file.
        Size:
          type: integer
          format: int64
          description: |-
            Total size in bytes — the sum of the manifest's **compressed** layer
            sizes, where Docker reports the uncompressed rootfs.
        VirtualSize:
          type: integer
          format: int64
          description: Deprecated alias of `Size`, kept for older clients.
...
    ImageSummaryResponse:
      type: object
      description: One entry of `GET /images/json`.
      required:
      - Id
      - ParentId
      - RepoTags
      - RepoDigests
      - Created
      - Size
      - SharedSize
      - VirtualSize
      - Containers
      properties:
        Containers:
          type: integer
          format: int64
          description: Containers using this image, `-1` when not counted.
        Created:
          type: integer
          format: int64
          description: Creation time, unix seconds.
        Id:
          type: string
          description: Image ID (`sha256:…`).
        Labels:
          type:
          - object
          - 'null'
          description: Labels; `null` when the image has none (Docker's shape).
          additionalProperties:
            type: string
          propertyNames:
            type: string
        ParentId:
          type: string
          description: Parent image ID.
        Platform:
          type:
          - string
          - 'null'
          description: Image platform, `os/arch` — **SatL extension**.
        RepoDigests:
          type: array
          items:
            type: string
          description: '`repository@digest` strings.'
        RepoTags:
          type: array
          items:
            type: string
          description: '`repository:tag` strings.'
        SharedSize:
          type: integer
          format: int64
          description: Bytes shared with other images.
        Size:
          type: integer
          format: int64
          description: On-disk size in bytes.
        VirtualSize:
          type: integer
          format: int64
          description: Deprecated alias of `Size`, kept for older clients.
...
    JsonErrorDetail:
      type: object
      description: '`errorDetail` of a [`JsonMessage`].'
      required:
      - message
      properties:
        message:
          type: string
          description: Failure detail.
    JsonMessage:
      type: object
      description: |-
        One line of a `POST /images/create` progress stream (Docker's
        `JSONMessage`).
      properties:
        error:
          type:
          - string
          - 'null'
          description: Fatal error message.
        errorDetail:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/JsonErrorDetail'
            description: Structured form of `error`.
        id:
          type:
          - string
          - 'null'
          description: Layer/blob the line is about.
        progress:
          type:
          - string
          - 'null'
          description: Pre-rendered progress bar.
        progressDetail:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/JsonProgressDetail'
            description: Byte counters.
        status:
          type:
          - string
          - 'null'
          description: Human-readable status.
    JsonProgressDetail:
      type: object
      description: '`progressDetail` of a [`JsonMessage`].'
      properties:
        current:
          type:
          - integer
          - 'null'
          format: int64
          description: Bytes transferred.
          minimum: 0
        total:
          type:
          - integer
          - 'null'
          format: int64
          description: Total bytes.
          minimum: 0
...